bugfix(#673): Dynamic cache creation PL bug
- Updated the data structures used so that it should be faster for large numbers
  of blocks, because hashtables are queried rather than the N^2 behavior of
  vectors.

+semver: bugfix

+++ include/fordyca/tasks/depth0/generalist.hpp
@@ -84,7 +84,8 @@ class generalist final : public foraging_task {
   void task_start(cta::taskable_argument* const) override {}
 
   rtypes::timestep current_time(void) const override RCSW_PURE;
-  rtypes::timestep interface_time_calc(uint, const rtypes::timestep&) override {
+  rtypes::timestep interface_time_calc(size_t,
+                                       const rtypes::timestep&) override {
     return rtypes::timestep(0);
   }
   void active_interface_update(int) override {}
