+++ src/tasks/depth2/cache_starter.cpp
@@ -61,7 +61,7 @@ double cache_starter::abort_prob_calc(void) {
 } /* abort_prob_calc() */
 
 rtypes::timestep cache_starter::interface_time_calc(
-    uint interface,
+    size_t interface,
     const rtypes::timestep& start_time) {
   ER_ASSERT(0 == interface, "Bad interface ID: %u", interface);
   return current_time() - start_time;
