+++ src/tasks/depth2/cache_transferer.cpp
@@ -62,7 +62,7 @@ double cache_transferer::abort_prob_calc(void) {
 } /* abort_prob_calc() */
 
 rtypes::timestep cache_transferer::interface_time_calc(
-    uint,
+    size_t,
     const rtypes::timestep& start_time) {
   return current_time() - start_time;
 } /* interface_time_calc() */
